Extract interposer placement cost logic into InterposerCostHandler
Move interposer crossing and congestion cost tracking out of NetCostHandler
into a dedicated class, and wire it through the placer and annealer.
