+++ vpr/src/place/annealer.h
@@ -17,6 +17,7 @@ struct t_placer_opts;
 enum class e_agent_state;
 
 class NocCostHandler;
+class InterposerCostHandler;
 class NetPinTimingInvalidator;
 class PlacerSetupSlacks;
 
@@ -183,6 +184,7 @@ class PlacementAnnealer {
                       const PlaceMacros& place_macros,
                       t_placer_costs& costs,
                       NetCostHandler& net_cost_handler,
+                      std::optional<InterposerCostHandler>& interposer_cost_handler,
                       std::optional<NocCostHandler>& noc_cost_handler,
                       const t_noc_opts& noc_opts,
                       vtr::RngContainer& rng,
@@ -293,6 +295,8 @@ class PlacementAnnealer {
     t_placer_costs& costs_;
     /// Computes bounding box for each cluster net
     NetCostHandler& net_cost_handler_;
+    /// Computes interposer cost terms when the device has interposer cuts
+    std::optional<InterposerCostHandler>& interposer_cost_handler_;
     /// Computes NoC-related cost terms when NoC optimization are enabled
     std::optional<NocCostHandler>& noc_cost_handler_;
     /// Contains weighting factors for NoC-related cost terms
